requirements.txt
pandas>=1.3.0
gspread>=5.0.0
openai>=1.0.0
konlpy>=0.6.0
numpy>=1.20.0
openpyxl>=3.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
faiss-cpu>=1.7.0
python-dotenv>=0.19.0